% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bndCMN}
\alias{bndCMN}
\title{bndCMN}
\format{
An \code{sf} object with the following fields:
\describe{
\item{\code{CMN}}{ ISTAT code for the District }
\item{\code{CMNd}}{ Name of the District }
}
}
\usage{
bndCMN
}
\description{
Digital Boundaries for Italian Municipalities
}
\keyword{datasets}
