% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bndPRV}
\alias{bndPRV}
\title{bndPRV}
\format{
An \code{sf} object with the following fields:
\describe{
\item{\code{PRV}}{ ISTAT code for the Province }
\item{\code{PRVd}}{ Name of the Province }
\item{\code{PRVs}}{ Road Code of the Province }
}
}
\usage{
bndPRV
}
\description{
Digital Boundaries for Italian Provinces
}
\keyword{datasets}
