% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{build_table}
\usage{
build_table(x, r)
}
\arguments{
\item{x}{the ISTAT code for the District}

\item{r}{the type of hazard: \code{hydro} or \code{geo}}
}
\value{
an htmlTable
}
\description{
Crete a tabnle to be used as label in leaflet map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
