% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dts}
\alias{dts}
\title{dataset}
\format{
A \code{data.table} object with the following fields:
\describe{
\item{\code{CMN}}{ Codice ISTAT del Comune }
\item{\code{var_id}}{ Identificativo della Variabile (chiave esterna per la tabella metadati \code{mdts}) }
\item{\code{value}}{ Valore assunto dalla Variabile sul Comune }
}
}
\usage{
dts
}
\description{
Tabella dati territoriali di rischio idrogeologico
}
\keyword{datasets}
