% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{empty_poly}
\alias{empty_poly}
\title{empty_poly}
\usage{
empty_poly(x = TRUE)
}
\arguments{
\item{x}{if \code{TRUE} the CRS of the polygon is the projected 3035 (ETRS89), otherwise the geodetic 4326 (WGS84)}
}
\description{
An empty polygon with CRS 3035 (\code{x = TRUE}, default) or 4326 (\code{x = FALSE})
}
