% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{extract_poly}
\alias{extract_poly}
\title{extract_poly}
\usage{
extract_poly(x)
}
\arguments{
\item{x}{an \code{sf} object}
}
\description{
Extract polygons from a geometry collection, or leave it as it is if not.
}
