% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroGeoRiskIT.R
\docType{package}
\name{HydroGeoRiskIT}
\alias{HydroGeoRiskIT}
\title{HydroGeoRiskIT}
\description{
Data and Digital Boundaries for the Analysis of Hazards and Risks from Floods and Landslides In Italy.
}
