% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_hydrogeo_risk.R
\docType{package}
\name{it_hydrogeo_risk}
\alias{it_hydrogeo_risk}
\title{it_hydrogeo_risk}
\description{
Data and Digital Boundaries for the Analysis of Hazards and Risks from Floods and Landslides In Italy.
}
