% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_risks.R
\name{map_risks}
\alias{map_risks}
\title{map_risks_cmn}
\usage{
map_risks(
  x,
  risk,
  neigh = FALSE,
  cmn = TRUE,
  cmnv = FALSE,
  with_mps = TRUE,
  add_tile = NULL,
  lvl_nms = NULL,
  lvl_cls = NULL,
  poly_tsp = 0.2,
  add_tbl = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{The ISTAT code(s) of the required District(s)}

\item{risk}{The Risk to draw on the map: \code{h}ydrological (floods) or \code{g}eological (landslides)}

\item{neigh}{IF \code{TRUE}, also includes all neighbours of the specified District(s)}

\item{cmn}{If \code{TRUE}, also draws the polygons of the District(s)}

\item{cmnv}{If \code{TRUE} (and also \code{cmn}is \code{TRUE}), adds the polygons of the neighbours of the District(s)}

\item{with_mps}{If \code{TRUE}, returns a complete map;
otherwise only the required polygons and the layers' menu}

\item{add_tile}{If \code{NULL} keep the default maptile \code{CartoDB.Positron},
otherwise a}

\item{lvl_nms}{If \code{NULL} keep the default names for the levels of the specified risk,
otherwise a character vector of the required length with the desired names}

\item{lvl_cls}{If \code{NULL} keep the default colour palette for the levels of the specified risk,
otherwise a character vector of the required length with the RGB values of the colours to be used}

\item{poly_tsp}{A value between 0 and 1 for the transparency of the polygons}

\item{add_tbl}{Add a table as label with information about people}

\item{verbose}{}

\item{cpoly_col}{}
}
\value{
un oggetto \code{leaflet}
}
\description{
Build a leaflet map for either the flood or landslide risk for one or more District,
including "neighbours".
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
