% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdts}
\alias{mdts}
\title{metadata}
\format{
Una data.table con i campi seguenti:
\describe{
\item{\code{var_id}}{ Identificativo della Variabile }
\item{\code{id_orig}}{ Identificativo della Variabile nel file ISPRA }
\item{\code{description}}{ Descrizione della Variabile }
\item{\code{label}}{ Etichetta della Variabile }
\item{\code{source}}{ Sorgente dati }
\item{\code{year}}{ Anno di misurazione }
\item{\code{risk}}{ Tipologia di Rischio }
\item{\code{level}}{ Descrizione del Livello di Rischio }
\item{\code{clevel}}{ Codice del Livello di Rischio }
\item{\code{measure}}{ Dato Territoriale di riferimento }
\item{\code{metric}}{ Specifica se Conteggio o Percentuale }
}
}
\usage{
mdts
}
\description{
Tabella di spiegazione delle variabili contenute nella tabella dati \code{dts}
}
\keyword{datasets}
