% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mps}
\alias{mps}
\title{mps}
\format{
An object of class \code{leaflet} (inherits from \code{htmlwidget}) of length 8.
}
\usage{
mps
}
\description{
\code{Leaflet} object to be used as initial map for shiny apps
}
\keyword{datasets}
