% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbours}
\alias{neighbours}
\title{neighbours}
\format{
A \code{data.table} object with the following fields:
\describe{
\item{\code{CMN}}{ ISTAT code for the District }
\item{\code{neigh}}{ ISTAT code for the bordering Districts }
}
}
\usage{
neighbours
}
\description{
For each Italian District, the list of all its bordering Districts
}
\keyword{datasets}
