% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{risks.lst}
\alias{risks.lst}
\title{risks.lst}
\format{
An object of class \code{list} of length 2.
}
\usage{
risks.lst
}
\description{
Description of the various Levels of Hazards and Risks
}
\keyword{datasets}
