% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoCMN}
\alias{geoCMN}
\title{geoCMN}
\format{
\describe{
\item{\code{CMN}}{ ISTAT code for the Municipality }
\item{\code{CMNd}}{ Name of the Municipality }
\item{\code{PRV}}{ ISTAT code for the Province }
\item{\code{PRVd}}{ Name of the Province }
\item{\code{PRVs}}{ Road Code of the Province }
}
}
\usage{
geoCMN
}
\value{
A \code{data.table} object with the following fields:
}
\description{
Dataset containing the complete list of Municipalities (as of 2022), and corresponding Province
}
\keyword{datasets}
