% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{palette.lst}
\alias{palette.lst}
\title{#' cmn.lst
#'
#' Complete List of Italian Municipalities (\code{CMN}), with Zip Codes and Provinces Roads Codes.
#'
#' @export
#'
cmn.lst <- create_cmn_lst()
palette.lst}
\format{
An object of class \code{list} of length 2.
}
\usage{
palette.lst
}
\description{
List of default colours to associate with the various Levels of Hazards and Risks
}
\keyword{datasets}
